requirements.txt
flask
sqlalchemy
twilio
python-dotenv
streamlit
pandas
plotly
streamlit-autorefresh
requests
